refactor: organize frontend by layer + aggregate (dogs)
