+++ frontend/src/api/dogs/getDogProfile.ts
