+++ frontend/src/components/dogs/DogProfileCard.tsx
