+++ frontend/test/api/dogs/editDogProfile.test.ts
@@ -1,4 +1,4 @@
-import { editDogProfile } from '@/api/editDogProfile';
+import { editDogProfile } from '@/api/dogs/editDogProfile';
 
 const dogId = 'a1b2c3d4-e5f6-7890-abcd-ef1234567890';
