+++ frontend/test/api/dogs/getDogProfile.test.ts
@@ -1,4 +1,4 @@
-import { getDogProfile } from '@/api/getDogProfile';
+import { getDogProfile } from '@/api/dogs/getDogProfile';
 
 const dogId = 'a1b2c3d4-e5f6-7890-abcd-ef1234567890';
